+++ R/vars-select.R
@@ -29,6 +29,7 @@
 #'   calling context.
 #' @param .include,.exclude Character vector of column names to always
 #'   include/exclude.
+#' @param .strict If `FALSE`, errors about unknown columns are ignored.
 #' @seealso [vars_pull()]
 #' @export
 #' @keywords internal
@@ -104,28 +105,32 @@
 #'   vars_select(names(mtcars), !! enquo(var1), !! enquo(var2))
 #' }
 #' wrapper(starts_with("d"), starts_with("c"))
-vars_select <- function(.vars, ..., .include = character(), .exclude = character()) {
+vars_select <- function(.vars, ...,
+                        .include = character(),
+                        .exclude = character(),
+                        .strict = TRUE) {
   quos <- quos(...)
 
-  if (is_empty(quos)) {
+  if (!.strict) {
+    quos <- ignore_unknown_symbols(.vars, quos)
+  }
+
+  ind_list <- vars_select_eval(.vars, quos)
+
+  # This takes care of NULL inputs and of ignored errors when
+  # `.strict` is FALSE
+  is_empty <- map_lgl(ind_list, is_null)
+  ind_list <- discard(ind_list, is_empty)
+  quos <- discard(quos, is_empty)
+
+  if (is_empty(ind_list)) {
     .vars <- setdiff(.include, .exclude)
     return(set_names(.vars, .vars))
   }
 
-  # Register vars to make them available to select helpers
-  scoped_vars(.vars)
-
   # if the first selector is exclusive (negative), start with all columns
   first <- f_rhs(quos[[1]])
   initial_case <- if (is_negated(first)) list(seq_along(.vars)) else integer(0)
-
-  # Evaluate symbols in an environment where columns are bound, but
-  # not calls (select helpers are scoped in the calling environment).
-  is_helper <- map_lgl(quos, quo_is_helper)
-  ind_list <- map_if(quos, is_helper, eval_tidy)
-
-  data <- set_names(as.list(seq_along(.vars)), .vars)
-  ind_list <- map_if(ind_list, !is_helper, eval_tidy, data)
 
   ind_list <- c(initial_case, ind_list)
   names(ind_list) <- c(names2(initial_case), names2(quos))
@@ -161,6 +166,48 @@ vars_select <- function(.vars, ..., .include = character(), .exclude = character
   }
 
   sel
+}
+
+ignore_unknown_symbols <- function(vars, quos) {
+  quos <- discard(quos, is_unknown_symbol, vars)
+  quos <- map_if(quos, is_concat_lang, lang_ignore_unknown_symbols, vars)
+  quos
+}
+lang_ignore_unknown_symbols <- function(quo, vars) {
+  expr <- get_expr(quo)
+
+  args <- lang_args(expr)
+  args <- discard(args, is_unknown_symbol, vars)
+  expr <- lang(node_car(expr), !!! args)
+
+  set_expr(quo, expr)
+}
+is_unknown_symbol <- function(quo, vars) {
+  expr <- get_expr(quo)
+
+  if (!is_symbol(expr)) {
+    return(FALSE)
+  }
+
+  !as_string(expr) %in% vars
+}
+is_concat_lang <- function(quo) {
+  quo_is_language(quo, quote(`c`))
+}
+
+vars_select_eval <- function(vars, quos) {
+  scoped_vars(vars)
+
+  # Symbols and calls to `:` and `c()` are evaluated with data in scope
+  is_helper <- map_lgl(quos, quo_is_helper)
+  data <- set_names(as.list(seq_along(vars)), vars)
+  ind_list <- map_if(quos, !is_helper, eval_tidy, data)
+
+  # All other calls are evaluated in the context only
+  # They are always evaluated strictly
+  ind_list <- map_if(ind_list, is_helper, eval_tidy)
+
+  ind_list
 }
 
 extract_expr <- function(expr) {