+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library("testthat")
+library("rlang")
 library("tidyselect")
 
 test_check("tidyselect")